ui-test-cleanup - move known-user into _lib/known-entities
